feat(mongoop.go): init mongo and try to establish db conn before server really start
the mongodb official go driver really implemented a connection pool and its related maintaining
procedure, we don't think we need to implement it ourselves, also, the cost of consideration about
the unstablitiy of network interchange is too high for us to do something to deal with that.
Finally, we chose just let it crash.