empty commit for note
The Anti-DDoS consideration of client side request forgery or consider
of database nosql injection may have some effect for anti-abuse.

But since we do need to check DB (if item exists or not), that will
always lead to server request to DB server. So this cost cannot be
reduced. Due to the above reason, I decided not to take any action
towards this.

If a user trying to query something but not exists or something wrong in
database, the server will either return HTTP 400/403/502 or a successful
response which the URI inside it is the index page.

The static page cost of explained response above can be safely ignored.

So we won't implement any further security check for Snippet ID due to
the consideration of both cost and effect.